requirements.txt
sounddevice
numpy
openai
scipy
python-dotenv